requirements.txt
pylangacq==0.15.0
PyYAML==5.4.1
pandas==1.3.0
numpy==1.20.0
tqdm==4.61.2
nltk==3.6.2
beautifulsoup4==4.10.0
librosa==0.8.1